Refactor: related #1
should refactor all ugly code, make it clearly and new featrues, doc..
