+++ bin/cli.js
@@ -2,97 +2,50 @@
 require('colors');
 const open = require('open');
 const keypress = require('keypress');
-const axios = require('axios').default;
 const R = require('ramda');
+const clearTerminal = require('../src/utils/clearTerminal');
+const streamList = require('../src/streamList');
+const {
+  createRequest,
+  createUserRequest,
+  createGamesRequest
+} = require('../src/request/createRequest');
 
-const TWITCH_CLIENT_ID = 'pl3x96lxedtdyvf3cmyai9fu4mod62';
-
-const createRequest = axios.create({
-  method: 'GET',
-  baseURL: 'https://api.twitch.tv/helix',
-  headers: {
-    'Client-ID': TWITCH_CLIENT_ID
-  }
-});
-const createUserRequest = axios.create({
-  method: 'GET',
-  baseURL: 'https://api.twitch.tv/kraken',
-  headers: {
-    Accept: 'application/vnd.twitchtv.v5+json',
-    'Client-ID': TWITCH_CLIENT_ID
-  }
-});
-const createGamesRequest = axios.create({
-  method: 'GET',
-  baseURL: 'https://api.twitch.tv/kraken/games/top',
-  headers: {
-    'Client-ID': TWITCH_CLIENT_ID
-  }
-});
-
-const clearTerminal = () => process.stdout.write('\u001B[2J\u001B[0;0f');
-clearTerminal();
-
-const streamList = [
-  {
-    _id: '493057',
-    name: "PLAYERUNKNOWN'S BATTLEGROUNDS - 絕地求生"
-  },
-  {
-    _id: '21779',
-    name: 'League of Legends - 英雄聯盟'
-  },
-  {
-    _id: '32399',
-    name: 'Counter-Strike: Global Offensive - 絕對武力'
-  },
-  {
-    _id: '138585',
-    name: 'Hearthstone - 爐石戰記'
-  },
-  {
-    _id: '29595',
-    name: 'Dota 2 - 鬥塔 2'
-  },
-  {
-    _id: '18122',
-    name: 'World of Warcraft - 魔獸世界'
-  },
-  {
-    _id: '494717',
-    name: 'IRL - 生活台'
-  }
-];
-let liveList = [];
-let normalizeListStreamList = [];
-
-const listLen = streamList.length;
+let liveList = [],
+  normalizeListStreamList = [];
 let currentSelectIndex = 0,
   currentShowLiveListSelectIndex = 0;
+const listLen = streamList.length;
+
 /**
  * selectMode
  * 1 => 尋找直播遊戲
  * 2 => 尋找台主
  */
-let selectMode = 1;
+const GAME_MODE = {
+  GAMELIST: 1,
+  LIVELIST: 2
+};
+let selectMode = GAME_MODE.GAMELIST;
+
+clearTerminal();
 
 keypress(process.stdin);
-
 process.stdin.on('keypress', (ch, key) => {
   clearTerminal();
   // console.log('got "keypress"', key);
 
   if (key && key.name === 'down') {
-    if (selectMode === 1) {
+    if (selectMode === GAME_MODE.GAMELIST) {
       if (currentSelectIndex + 1 > listLen - 1) {
         currentSelectIndex = 0;
       } else {
         currentSelectIndex = currentSelectIndex + 1;
       }
       show();
-    } else if (selectMode === 2) {
+    } else if (selectMode === GAME_MODE.LIVELIST) {
       let normalizeListStreamListLen = normalizeListStreamList.length;
-      if (currentShowLiveListSelectIndex + 1 > normalizeListStreamListLen) {
+      if (currentShowLiveListSelectIndex + 1 > normalizeListStreamListLen - 1) {
         currentShowLiveListSelectIndex = 0;
       } else {
         currentShowLiveListSelectIndex = currentShowLiveListSelectIndex + 1;
@@ -101,14 +54,14 @@ process.stdin.on('keypress', (ch, key) => {
     }
   }
   if (key && key.name === 'up') {
-    if (selectMode === 1) {
+    if (selectMode === GAME_MODE.GAMELIST) {
       if (currentSelectIndex - 1 < 0) {
         currentSelectIndex = listLen - 1;
       } else {
         currentSelectIndex = currentSelectIndex - 1;
       }
       show();
-    } else if (selectMode === 2) {
+    } else if (selectMode === GAME_MODE.LIVELIST) {
       let normalizeListStreamListLen = normalizeListStreamList.length;
       if (currentShowLiveListSelectIndex - 1 < 0) {
         currentShowLiveListSelectIndex = normalizeListStreamListLen - 1;
@@ -118,13 +71,13 @@ process.stdin.on('keypress', (ch, key) => {
       showLiveList(normalizeListStreamList);
     }
   }
-  if (key && key.name === 'return') {
-    if (selectMode === 1) {
+  if (key && (key.name === 'return' || key.name === 'right')) {
+    if (selectMode === GAME_MODE.GAMELIST) {
       fetchLiveStream().then(liveStreamResponse => {
         /**
          * selectMode 2
          */
-        selectMode = 2;
+        selectMode = GAME_MODE.LIVELIST;
         liveList = liveStreamResponse;
         normalizeListStreamList = normalizeLiveStream(liveList);
         clearTerminal();
@@ -139,6 +92,11 @@ process.stdin.on('keypress', (ch, key) => {
         }
       );
     }
+  }
+  if (key && key.name === 'left') {
+    selectMode = GAME_MODE.GAMELIST;
+    clearTerminal();
+    show();
   }
   if (key && key.ctrl && key.name === 'c') {
     process.stdin.pause();
