+++ src/streamList.js
@@ -1,0 +1,32 @@
+const streamList = [
+  {
+    _id: '493057',
+    name: "PLAYERUNKNOWN'S BATTLEGROUNDS - 絕地求生"
+  },
+  {
+    _id: '21779',
+    name: 'League of Legends - 英雄聯盟'
+  },
+  {
+    _id: '32399',
+    name: 'Counter-Strike: Global Offensive - 絕對武力'
+  },
+  {
+    _id: '138585',
+    name: 'Hearthstone - 爐石戰記'
+  },
+  {
+    _id: '29595',
+    name: 'Dota 2 - 鬥塔 2'
+  },
+  {
+    _id: '18122',
+    name: 'World of Warcraft - 魔獸世界'
+  },
+  {
+    _id: '494717',
+    name: 'IRL - 生活台'
+  }
+];
+
+module.exports = streamList;
