+++ src/utils/clearTerminal.js
@@ -1,0 +1,3 @@
+const clearTerminal = () => process.stdout.write('\u001B[2J\u001B[0;0f');
+
+module.exports = clearTerminal;
